Add files via upload
This project is designed to predict the potability of water—whether it is safe for human consumption—by analyzing various water quality parameters. Water quality is a critical factor for public health, and developing a predictive model can help authorities and individuals monitor and ensure safe drinking water. The project uses machine learning to identify patterns and relationships in the data, providing an efficient and automated solution for water quality assessment.